Separate files for FelicaHelper class
