+++ FelicaLib_Remodeled/FelicaLib_DotNet/Properties/AssemblyInfo.cs
@@ -35,7 +35,7 @@ using System.Runtime.InteropServices;
 // すべての値を指定するか、下のように '*' を使ってビルドおよびリビジョン番号を 
 // 既定値にすることができます:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.2.45.0")]
-[assembly: AssemblyFileVersion("1.2.45.0")]
+[assembly: AssemblyVersion("1.2.46.0")]
+[assembly: AssemblyFileVersion("1.2.46.0")]
 
 [assembly: CLSCompliant(true)]
